Create index.html
Upload of the Livelens.site index file. Note the style line  in the head is commented out, this is because of the layout of the server. Style files are inside of a public directory, but when developing on local host this does not work. To work on this file on a local machine uncomment that line and place the style sheet file in the same directory.